I forgot what I did here but it was something important.

+++ src/Projectile.java
@@ -1,7 +1,7 @@
 import java.awt.*;
 
 public class Projectile extends Sprite{
-    public Projectile(Image i, int x, int y, int width, int height, int speedX, int speedY) {
+    public Projectile(Image i, int x, int y, int width, int height, double speedX, double speedY) {
         super(i, x, y, width, height, speedX, speedY);
     }
 }
