+++ src/Sprite.java
@@ -11,7 +11,7 @@ public class Sprite {
     protected double speedX;
     protected double speedY;
 
-    public Sprite(Image i, int x, int y, int width, int height, int speedX, int speedY) {
+    public Sprite(Image i, int x, int y, int width, int height, double speedX, double speedY) {
         this.i = i;
         this.width = width;
         this.height = height;
@@ -53,5 +53,6 @@ public class Sprite {
     public void setY(double y) {
         this.y = y;
     }
+
 }
